requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=4.1.0
flask>=2.3.0
twilio>=8.10.0
reportlab>=4.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0 